+++ packages/stellar-axelar-std/src/lib.rs
@@ -4,6 +4,9 @@
 #[cfg(any(test, feature = "testutils"))]
 extern crate std;
 
+#[cfg(feature = "alloc")]
+extern crate alloc;
+
 #[cfg(any(test, feature = "testutils"))]
 pub mod testutils;
 
@@ -11,6 +14,8 @@ pub mod testutils;
 mod tests;
 
 pub mod traits;
+
+pub mod string;
 
 pub mod types;
